test-eswitch-add-in-mode1-del-in-mode2.sh: Fix REP dev name issue
The rule added for vxlan is added to the REP device, though instead of
using the device name it used the NIC and inferred the rep name. This
method is not compatible with platforms that name the representors
according to a different format, and therefore the REP device name is
taken from the configuration script.

Signed-off-by: Gavi Teitz <gavi@mellanox.com>