common.sh: Print errors to stdout
This is because mars log showing stderr output after stdout
and makes it hard to read which failures belong to which case.

Failures with fail() command still prints to stderr because they
are used inside functions that can return value using stdout.
Also fail() command stops the test so it will always be the last error
of the last run case.

Signed-off-by: Roi Dayan <roid@mellanox.com>