common.sh: Sync journalctl data at the end of the test
Sync all pending log before checking for errors. this is to make
sure we go over all the log collected during the test.

Need to consider adding it before each call to journalctl for checking
the log but also need to check and consider run time.

With old journalctl that doesn't support sync we sleep half a second instead.

Signed-off-by: Roi Dayan <roid@mellanox.com>