tc_iperf_common.sh: Verify that iperf3 is installed
Currently, if iperf3 is not installed, test-tc-traffic* fail
without providing meaningful error message to the user:

ERROR: Zero packets on filter

Verify that iperf3 is installed at the beginning of tc_iperf_common.sh
script. Exit the test and properly notify the user if no iperf3 binary is
found on the system.

issue: 1821539
Change-Id: I875922816b58be03aee9965bd4edf6770dc91874
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>