test-tc-insertion-rate-skip_hw.sh: Add new test
Add new test wrapper for test-tc-perf-update.sh that compares benchmark
results with results saved in persistent file. If persistent file is not
present, create it with default or user-provided name and populate it with
reference results from current execution. Put all infrastructure code into
tc_tests_common.sh header in order for it to be reused by similar test for
fw and sw steering in following commits.

issue: 1911445
Change-Id: I7250977f3c3e2c298ec7b4f50fa35feb4f501af8
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>