test-tc-rule-act-add.sh: Change actions used for test
The test verifies all existing action types, including the ones that are
not supported for offloading by mlx5 and are not compiled by default by
distros and out configs used to build test kernels. As such, breakage for
these actions are not caught by regression since test is designed to skip
testing any action modules that were not compiled.

Remove actions that are not supported by mlx5. Implement support for ct
action that was added after the test.

Change-Id: Ia10ea21e705b0d297faebd1e37c0808600e0bb43
Signed-off-by: Vlad Buslov <vladbu@mellanox.com>