test-all.py: Add html format for test logs
Since we already have flag for html format that currently save
summary.html users would expect the summary to include links to test
logs in html format as well, but current implementation is linking to
ansi color log files which is unreadable in web browsers.
This commit is adding html log file in case the --html flag is given and
updating the links in the summary.html to use the html files and not the
log files.

Change-Id: Ie2b39d6221e5cabba3123dfce00ec26c1e17cd1e
Signed-off-by: Guy Ergas <guye@mellanox.com>