delete test test-vf-lag-sriov-disable.sh
As it is a FW issue that needs a wide ARCH change,
it was decided to not implement it in the near future,
it will be published as RN as unsupported flow.

The only change is that the FW implented is to return
syndrome in such case instead of silent failure.

Issue: 2108183
Issue: 2035950
Change-Id: I619a143e01d1379e7fea7261629162f735d04976
Signed-off-by: Raed Salem <raeds@mellanox.com>