common.sh: Wait for more time when switching mode with ofed compat
In ofed compat we change mode in async way and we sleep.
Sometimes we need to sleep more than even 20 seconds. i.e. 127 reps.
Calc according to number of vfs because we dont want staticly to sleep
long time so if we did fail i.e with just 10 reps we don't want to sleep
too much before deciding this is an issue.

Change-Id: I3460be2fd64de51e24c1c750c1d8ead0ee8b3b63
Signed-off-by: Roi Dayan <roid@nvidia.com>