common-ecmp.sh: Sleep 0.5s before checking nics exists
If port2 was in switchdev and common.sh disabled it in the start
of the test it will take half a second for the interface to appear again.
This is if nic and uplink rep are not the same netdev (uplink rep redesign feature).
Usually tests do stuff before checking for nics. like change mode and that function
wait for nics so the sleep only needed the the ecmp tests that right away check for nics.

Change-Id: Id71f6f1d03eeba1fff23549fa9a4c4289b7f1286
Signed-off-by: Roi Dayan <roid@nvidia.com>