Revert "test-tc-wrong-geneve.sh: fix print"
This reverts commit 92109cf16e0ae5540e92e5b45422ea86a2fe9d58.

Dima Chumak

This fix looks wrong. Unfortunately BASH syntax for array slicing isn't intuitive and can be confusing:

    ${encap_ids[@]:0:1}

The last field after ':' is length and not an array element index. So, with this patch the 'encap IDs 1st-2nd:' log shows only single ID:

    encap IDs 1st-2nd: 0x0
    encap IDs 2nd-3rd: 0x0 0x0

It's not easy to find where this syntax is explained in BASH man page either, probably I should have left a comment about it in the test script:

    man bash | less +/parameter:offset:length

I think it's better to revert this change.

Change-Id: Ia5c5a4e1ab7f0d1957459f0dcab5cbad1117d424
Signed-off-by: Roi Dayan <roid@nvidia.com>