common.sh: Add functon on_remote_dt
This function can be used to execute other commands defined in
common.sh on a remote host.

Useful for running commonly used commands on remote host without having
to rewrite.

It assumes that the common.sh is accessible at the same location on remote
as on the local host (caller).

Change-Id: I1a0b82f7658ef20d1d2c4e851c360cb5798bb289
Signed-off-by: Alaa Hleihel <alaa@nvidia.com>