test-ovs-ct-vxlan.sh: Run iperf daemon local and client on remote
We sometimes have an issue in CI with this test that the server
started later than the client. either ssh conn took time or daemon
to start took time.
This change removes the needed sleep between daemon and client as
the ssh command wont finish until iperf daemon starts and no need
for another ssh command later to kill the daemon.

Change-Id: Ib25715948ba110563c29333747c1da17b8ae2747
Signed-off-by: Roi Dayan <roid@nvidia.com>