common: Fix ovs-dpdk initialization arguments
Any DPDK application requires as an argument the devices which DPDK
should bind to prior to starting the application, if no devices are
specified using the whitelist argument DPDK will attempt to probe
all devices on the system, in case a device is not DPDK compatibale
then DPDK will seg fault and crash.

When working with ovs-dpdk dpdk-init flag should be turned on which
indicates start of the DPDK application.
prior to setting dpdk-init flag to true, dpdk-extra field should be
filled with the wanted devices to bind to DPDK.

In case dpdk-extra is not defined then setting dpdk-init to true
will attempt to probe all the devices on the system due to default
DPDK behavior.

Make sure to always set dpdk-extra in case not set before enabling
dpdk.

Change-Id: I61c3f294a02373d46673cc1972a596ff87c5fdf0
Signed-off-by: Salem Sol <salems@nvidia.com>