ipsec-tests: Fix expected to see packets over UDP when running test-ipsec-crypto.sh
currently running test-ipsec-crypto.sh in stage of test ipsec in transport
mode with 128 key length using ipv4 with offloaded TX the test passes with
tcp traffic but fails with UDP traffic. The issue is caused because of a
tcpdump udp filtering misuse when ESP packets are expected.
Fix by removing the filter

Issue: 2851582
Change-Id: I374b777cf4e97ca941eafddc89b56d54ee56f4e7
Signed-off-by: Emeel Hakim <ehakim@nvidia.com>