ipsec-tests: Apply a fix to test-ipsec-crypto-ipv6.sh
commit 197c9b58 fixed a test bug in test-ipsec-crypto.sh where running
test-ipsec-crypto.sh in stage of test ipsec in transport mode with 128
key length using ipv4 with offloaded TX the test passes with tcp traffic
but fails with UDP traffic. The issue is caused because of a tcpdump udp
filtering misuse when ESP packets are expected.
Fix by removing the filter.
this commit applies the same fix to test-ipsec-crypto-ipv6.sh

Issue: 2851582
Change-Id: Ibc4c4225f12a79ded411fccc5f5b4f6fdee04995
Signed-off-by: Emeel Hakim <ehakim@nvidia.com>