ovs-dpdk-tests: Fix test-ovs-dpdk-vxlan-ct-udp.sh
Due to a bug in iperf3 [1] which sends TCP packet even when specifying UDP
test-ovs-dpdk-vxlan-ct-udp falsely passes since the TCP traffic gets dropped due to
the OpenFlow configuration in this test leading to the iperf3 connection not
getting established thus no traffic passes. Revert back to iperf, and fix
sleeping times in the test.

[1] https://serverfault.com/questions/808918/iperf3-uses-tcp-even-if-udp-is-specified

Issue: 2866324
Change-Id: I80c5f8c09dbdafe32c4c5b6d95efb2d95842cfe3
Signed-off-by: Salem Sol <salems@nvidia.com>