testpmds tests: Set value total-num-mbufs to 4gb when running testpmd
OVS project increased max port from 256 to 1024 due to the need to
support 512 interfaces. This change affected DPDK package default
values.

DPDK testpmd application has its own logic to calculate the buffers and
memory to be used, with this change, the binary testpmd in OFED needs to
have ~4 more times hugepage space for mbuf. Or it could be simply got
rid of by using "--total-num-mbufs" to limit the memory usage.

Issue: 2862699
Change-Id: I5080e4a121dd29faabc57d3ff0e0b7cb29336382
Signed-off-by: Mahmoud AbuZant <mahmoudabuz@nvidia.com>