ct-udp: Wait 3 seconds to check ct udp connections
After a upstream commit, udp connections are offloaded 2 seconds later
after the connections are established. Currently, the ct udp tests only
wait 2 seconds to check if the ct udp connections are offloaded or not.
Now it is not enough. Change it to 3 seconds.
b7b1d02fc439 ("netfilter: conntrack: set on IPS_ASSURED if flows enters internal stream state")

Issue: 2852861
Change-Id: I0c2280b190967411ff652fc30f7d2bc3e567ddcd
Signed-off-by: Chris Mi <cmi@nvidia.com>