test-ovs-ct-vf-lag: Avoid reconfig during traffic
This check is redundant currently and fails in debug kernels as
the system is slow and busy with iperf packets to offload new rules
quickly. even the 6 seconds are not enough. remove that operation.

While editing the tests use /proc/net/nf_conntrack instead of conntrack -L
as we can see HW_OFFLOAD instead of assured.

Issue: 2901389
Change-Id: I478f64fc76379dacef1f19100e464a7c8a4a68f4
Signed-off-by: Roi Dayan <roid@nvidia.com>