common-dpdk: Refactor connection offload checker grep
Currently to check for connections offload we grep "Total  CT bi-dir Connections:"
this grep is highly sensitive for the amount of spaces between Total and CT, in
newer ovs versions these amount of spaces increases which will lead the grep to
fail. Instead, split the grep into two, one for "Total"
and the other for "CT bi-dir Connections:" for more robustness.

Issue: 3077223
Change-Id: Ib4f767e5c1497b833dd58c315b8673eec01ccae7
Signed-off-by: Salem Sol <salems@nvidia.com>