setup_configure: Remove calling virsh undefine for vdpa setups
In order to run vDPA setups we need to change the VM XML file to include the
vhostuser ports, this was done by first undefining the VM then editing the XML
and finally re-define it. This way leads to losing all the VM data and saved
files which we need if we want to keep the ssh keys created prior to editing
persistent. Instead, do not undefine the VM in order to keep the ssh keys created
upon running configure_asap_devtest_env.sh.

This is safe to do since virsh define can be used without the need to undefine
first.

Issue: 3112870
Change-Id: Ie47bf5440caa2821de41e7e2b7c0229f633f2552
Signed-off-by: Salem Sol <salems@nvidia.com>