ovs-dpdk-tests: common-dpdk.sh: Initialize DPDK with dummy PCI
DPDK stores caches for ports upon probe and attaching the port, removing and
re-adding a port might cause it to be loaded with the cached info. This might
cause problem for ovs.
A workaround for this can be to either restart openvswitch
when removing a port to force DPDK to flush its caches or re-probe each device
when attaching it.
The better solution is to re-probe the devices upon attaching, and in order to
do so we'd need to start ovs with a dummy, non-existing PCI, then adding the
needed devargs per port. Do it by changing dpdk-extra arguments from taking
the PF PCI, it representors and other dpdk args to probe a dummy PCI and move
the arguments to the options upon adding a port to ovs bridge.

Issue: 3128901
Change-Id: I20bad6da0cfdb831b8517fa8f6f058a5f35dc5ad
Signed-off-by: Salem Sol <salems@nvidia.com>