test-ovs-dpdk-ct-after-nat.sh: Fix test
Currently the test adds VF2 to ns0 and VF to ns1, this goes against the assumptions
taken for VDPA setups in which VF is added to ns0 and VF2 to ns1 and causes the test
to fail. Fix it by aliging the test to the standard VF to namespace assigning
done in all other tests.

Also, the openflow rules added in this test do not take into consideration icmp
traffic. Add a rule for it.

Issue: 3123196
Change-Id: Iaa8056e95c39344fe7ed638d70bc55bb9a8a0860
Signed-off-by: Salem Sol <salems@nvidia.com>