common.sh: Add connectin timeout log print
This will help check logs later if tests fail because
remote configuration fails but test didn't check rc
and printed an error.

The rc for ssh connection timeout is 255.
If later we see commands returning the same error code
need to change/update those commands or we will have
to remove the print.

Change-Id: I084b5ccaefc8c5b8c5f3fcb4d408ed5098de9151
Signed-off-by: Roi Dayan <roid@nvidia.com>