test-eswitch-ingress-rate-limit.sh: fix cx6dx compatibility
The test has been disabled due to bug for more than a year and during that
time got incompatible in several places:

- Test already ignores first second of iperf3 traffic when calculating
total rate due to the fact that after eswitch enable it takes some time to
stabilize the vqos tree. However, 1 second seems to be not enough for cx6dx
with current stable firmware. Instead of trying to tune the timeout add
dedicated iperf3 warm-up run after configuring the target net device.

- With introduction of internal port offload the traffic between VF and OvS
bridge netdevice is now offloaded. However, the offload doesn't work when
uplink link is down. Disable this part of the test on new NICs since that
is not intended configuration for internal port offload and functionality
it well-tested by VF-to-VF traffic anyway.

- For whatever reason REP2 is down by default on cx6dx after enabling
SRIOV. Manually set the link up.

- Unrelated to cx6dx test sometimes fails on all NICs due to rate being
slightly out of limit range. To prevent this rare but annoying failure
increase test rates to 16 and 32 mbps. This increases the tolerances (which
are calculated as percentage of the value) and also HW rate limiting is
reported to be more stable on higher rates.

Issue: 2790606
Change-Id: I4dfb58c785b0eb94e60ba9b384db7fe81bde22bc
Signed-off-by: Vlad Buslov <vladbu@nvidia.com>