common-ipsec-offload.sh: Fix crypto tests running sometimes in legacy
Usuaully a default startup for the tests, the ports are in switchdev mode
but after a trusted vfs test the ports left in legacy mode,
so crypto tests could run in inconsistent state, either in switchdev or legacy mode.
There is no config_cyrpto func called in all tests but there is
cleanup_crypto that always called in the start of the test and before
set trusted vfs in such tests.
Add a call to set ports to switchdev in cleanup_crypto so tests are consistent.

Change-Id: I07a906a6d1afb5964a69f8e36e12d9f46d849f7c
Signed-off-by: Roi Dayan <roid@nvidia.com>