config sriov if needed before enabling switchdev
The below tests are failing to find reps after enabling switchdev.
The issue is due to these tests are not enabling sriov if needed 
before moving to switchdev and it seems one of the previous tests 
is disabling sriov without enabling it back so we hit this issue.

Failing Tests:
- test-ovs-ct-scapy-udp.sh
- test-ovs-ct-scapy-tcp-aging.sh
- test-ovs-ct-scapy-hw-ovs-miss.sh

Issue: 3185801
Change-Id: Ia5af46ba0476ff698f67dc69affaba25453de994
Signed-off-by: Ziyad Atiyyeh <ziyadat@nvidia.com>