test-tc-vxlan-meter.sh: Fix find rate error
We get rate from the 6th field of the iperf summary. The line is like:
[SUM]  0.0-10.0 sec  243780600 Bytes  24401131 Bytes/sec   0.078 ms 237646/411775 (58%)

However, iperf sometimes doesn't run 10 seconds, and the line is like:
[SUM]  0.0- 9.9 sec  242384800 Bytes  24364403 Bytes/sec   0.011 ms 224173/397305 (56%)
Now the 6th field of this line is "Bytes", not the bandwidth. To fix
issue, reading the bandwidth starting from the end of the line.

Besides, as we don't use offloading in the other (client) side, it's possible
that client can't generate enough traffic. If the traffic is almost the target
bandwidth, server may get lower number than expected. So, lower the target rate
to avoid this failure.

Issue: 3236982
Change-Id: I9957f75ea79f4a1ec34139d5d4880ea6223084e7
Signed-off-by: Jianbo Liu <jianbol@nvidia.com>