test-bond-pf-to-bridge.sh: Properly fix bond initialization
Due to recent changes in mlx5 that preserve the mode even when setting
number of VFs to zero previous fix only worked when the DUT was in
switchdev mode before running the test. However, since the test is intended
to be run in legacy where relevant dmesg log is not generated when bond is
initialized, it started failing in the intended configuration.

Fix the test:

- Ensure DUT is in legacy mode by calling disable_sriov() in config function.

- Revert to calling __config_bonding() since config_bonding() fails in
legacy due to missing dmesg log.

- Since there is no good way to ensure LAG is fully initialized in legacy
mode (recent debugfs addition is not present on any LTS kernels), just
sleep for 3 seconds after calling __config_bonding().

Issue: 3257403
Change-Id: I8c5edd3e8984f1182c03596a04608b9f9caabaf0
Signed-off-by: Vlad Buslov <vladbu@nvidia.com>