test-ct-meter-udp.sh: Add another meter with half rate
There is a backporting issue in OFED 5.8, the police action index is
not passed correctly to driver for old kernels. Then driver uses the
default value (0) as the index, and all the police actions share the
same hardware meter which is wrong.
To detect this issue, add another meter with half rate in this tests.

Issue: 3232445
Change-Id: Icd4be8d00ab0a8f72361adc4ac7af29755c1ef5c
Signed-off-by: Jianbo Liu <jianbol@nvidia.com>