Make config_multipath_route() also check vf lag state
vf lag state is being tested after every call to config_multipath_route()
so just to the call inside.
This fixes the cases missing the check.
Add in test-ecmp-add-multipath-rule.sh a clean in first case so it
won't fail next case that expected ecmp to be or not be configured.

Change-Id: Ie3597570f83bc453c18849f5efecb106c6db6e04
Signed-off-by: Roi Dayan <roid@nvidia.com>