udp-perf: Interval enhancement
previous udp-perf is limited to send max 256 packets,
and 10 packets in sec, the new udp-perf.py enhanced to send more packets.
and added interval to send packets faster

in common-ovn we increase the packets we send to 15000 (timeout * 1000).
and 1000 packets per second ( --inetval=1000 ).

in addition adapted this format in ICMP traffic as well. (15000 packets, 1000 packets per sec)

this will help to send more more packets to sufficient buffer to pass in SW.

Issue: 3352392
Signed-off-by: Omri Barazi <obarazi@nvidia.com>
Change-Id: Id5401dd4600480e002467b841802d60e5b797b37