macsec-tests: Expand xpn tests to cover one more scenario
The macsec xpn feature triggers an object change event each time
the packet number (pn) is a multiply of 2^31, currently the test
checks only the case of 2^32 which dont have same behavior as 2^31
case.

Extends the test to cover the other case of 2^31 to have a full
coverage for the feature.

Change-Id: I27d025ef411ae5a93a282d0b2b042993d645edfd
Signed-off-by: Emeel Hakim <ehakim@nvidia.com>