test-mpls-over-udp-traffic.sh: Add skip_hw for tc rules on remote
As we start to check error on remote server, this test fails because
remote fails to parse tunnel attributes. The reason is that, unlike
local, flexparser is not enabled in the remote's firmware.

For this test, both traffic and the offloaded rules are already
verified on local, it's not necessary to verify the offloading on
remote. So add skip_hw on remote to resolve this issue.

Issue: 3439339
Change-Id: I36507bd33ce5e0797cefdd6e3186823e03e3bedd
Signed-off-by: Jianbo Liu <jianbol@nvidia.com>