Revert "common.sh: Remove DBG line prints from dump_ovs_log"
This reverts commit d4316364d693b4a5bb81ec6ebec6be8fac08adb6.

If a test fails in CI, the log is not provided, but only this dump.
Fail to offload is not an error or a warning, thus there is no way to
tell what failed to be offloaded, or why in such cases.
Enable back the DBG prints to provide such visibility.

Issue: 3453302
Change-Id: Ib7a509cf939d35bcbbad5ae187078099f1f390e3
Signed-off-by: Eli Britstein <elibr@nvidia.com>