test-tc-vxlan-gbp.sh: Add dp rules check in traffic test
To check vxlan traffic is offloaded correctly is not enough for vxlan gbp.
TC module may silently drop the GBP options if the kernel doesn't support
them, so the rules may be offloaded as plain VxLAN rules instead of VxLAN
GBP ones.

Add new check that tc filter was created with vxlan gbp action or match.

Change-Id: Icaafc6374e66f9108892a71778db296e2f86ab90
Signed-off-by: Gavin Li <gavinl@nvidia.com>