test-ct-meter-udp.sh: Refactor the test to make it more robust
Fix several issues with the test that make it prone to arbitrary failures
in regression:

- Set rate limit to 100Mbit/s. Cloud setups with full debug kernel often
can't handle 200Mbit/s traffic generation which might result receiver
traffic rate to fall below 180Mbit/s and consecutively result the test
failure.

- Use iperf3 server '1' option to terminate the server after one traffic
run. Without this iperf3 server might generate multiple JSON objects in its
output due to stray UDP packets arriving after the session termination
which results errors with jq.

- Use iperf3 'O' option to omit first 3 seconds of traffic from the total
rate calculation. This is necessary because CT UDP flows may take from
tens of milliseconds (with UDP NEW offload) to two seconds (without UDP NEW
offload) to offload a UDP connection. During that period deviations in rate
limiting are expected. Update jq to skip intervals marked as 'omitted' when
calculating the average rate.

- Current verify_iperf3_bw() implementation just uses a second sample from
the intervals range to measure the rate instead of total run rate (probably
as a workaround for longstanding iperf3 bug that outputs
.end.sum_received.bits_per_second as 0 for UDP flows). Not only this is
quite imprecise, but also the worst period for sampling as we expect
fluctuations in metering when flows are promoted from SW to HW offload.
Manually calculate the average rate with jq magic. When doing so ignore
last interval which is usually some fraction of a second and can deviate
significantly from other full one-second periods (which also could be
solved by extending the iperf3 runtime in order to prevent a single
interval from skewing the average but will undesirably increase the
regression runtime).

- Sleep for 2 seconds instead of 0.5 seconds after starting iperf3 server
since it might not be fully initialized after half a second on full debug
kernels. Also increase iperf3 timeout values to be more generous.

- Remove bogus comment (probably result of copy/paste) that states that CT
connection is expected to be offloaded after first 4 packets which is only
the case for TCP and not for UDP.

Issue: 3195822
Change-Id: If16864da5c58e3f951f38b805e5a8d115b6128a2
Signed-off-by: Vlad Buslov <vladbu@nvidia.com>