setup_configure.py: Disable flow control on PF
This is a workaround for a current issue with vf lag and dpdk.
When adding a port to ovs we disable flow control on it as
we dont config ovs to do flow control.
If a single pf test is running and after it a vf lag test is
running we are in a situation where pf0 has disabled flow control
while pf1 has it enabled and using vf lag with dpdk cause a segfault.
To WA this disable flow control when configuring the system.

Issue: 3601477
Change-Id: I3cd722a615a69e9223b2e440702041f118615c51
Signed-off-by: Roi Dayan <roid@nvidia.com>