common.sh: Add more words to ovs log checker
Separate real err words from normal ERR tagged lines as
sometimes that are not real errors.
i.e.
2023-10-01T16:45:49.533Z|00003|netdev_offload_tc(revalidator13)|ERR|dump_create: failed to get
ifindex for ovs-system: No such device

ovs-system could not exists and it's not an error. i.e. in ovs-dpdk configuration.
need to fix this this in ovs not to print an error.

Change-Id: I9f7a5e0a07574a9a2ac7ea44dce014f08ba74caa
Signed-off-by: Roi Dayan <roid@nvidia.com>