common-ovn: Add sleep before checking for dpdk offloads
In some cases iperf traffic is finishing and we immidiately start checking for
offloads this results in having the tcp teardown stream iperf3 sends to appear
as not offloaded however the rule gets offloaded but we're racing between the
offload request and the query of the rule. To avoid this race sleep for 0.5
seconds before checking for offloads.

Issue: 3655968
Change-Id: Id70953365fd17b5500a5ebe461ec176d6d440b0e
Signed-off-by: Salem Sol <salems@nvidia.com>