test-ovs-dpdk-openflow-rules-insertion.sh: Improve openflow rules insertion rate
The pmd threads and vswitch threads accessing the openflow table
causing RCU contention. To make sure that rcu synchronizes on idle
configure pmd-quiet-idle to be true to allow the PMD threads to
go into quiescent mode when idling.

An example result with CX6DX
before the commit
4000 rules took 6668 ms.

after the commit
4000 rules took 384 ms.

Issue: 3659033
Signed-off-by: Mahmoud AbuZant <mahmoudabuz@nvidia.com>
Change-Id: Ia09da15c4d3efd0620daca53f721e84dc0299753