test-all.py: Ignore min_kernel for BF setup
In BF setup we have a single kernel and the mlnx ofed
is the driver but min_kernel variants are compared against
the host kernel.
I.e. host kernel rhel8.2 could have old kernel but it doesn't
affect the features supported inside the BF ARM.
Ignore min_kernel variants when on BF setup so they are
being compared only when working in BF nic mode.

Change-Id: I2a113f4232ff0f8d9e4471119508192a2ca96d56
Signed-off-by: Roi Dayan <roid@nvidia.com>