common-devlink.sh: Fix to newer SF interface
The common-devlink.sh file is old and expect an old
interface when SFs were first introduced.

The two tests using this commit were disabled for
a year and changes were not detected.

Instead of full modification of the tests to use common-sf.sh
update what is needed and cleanup later.

Change-Id: I04403e78e8407b94b785488ec9fbaacc745a7135
Signed-off-by: Roi Dayan <roid@nvidia.com>