common-dpdk.sh: Fix check_dpdk_offloads
With recirc rules and tunnels it could be some rules after recirc 0x0
are matching the tunnel ip and not the inner ip.
The function filtered those rules out.
This is an intermediate fix to catch tests false passing.
A better will be to remove matching of $IP entirely.
All rules should be offloaded, filtering out rules with unsupported match or action only
and not to create a filtered list of only rules with specific IP.

Change-Id: If2ab058e9eb58efe0b634d0162c9a23c268f1d4a
Signed-off-by: Roi Dayan <roid@nvidia.com>