test-ovs-dpdk-openflow-rules-insertion.sh: Relocate configuration of 2 VFs to the cleanup phase
When unbinding a device from vfio or uio before hotplug removal
and subsequently rebinding it with a different configuration,
dpdk may generate an "Unexpected device scan at <path>" error.
To address this issue, this commit relocates the configuration
of 2 VFs to the cleanup phase, which is the appropriate
location for such operations.

Issue: 3717492
Change-Id: I19357078f78185c4264b244672c60b2a70f5a8ed
Signed-off-by: Emeel Hakim <ehakim@nvidia.com>