common.sh: Fix comparison with empty RM status
After the cited commit we hit issue of missing quotes
comparison with empty var and also empty warnings on RM
info that doesn't exists as RM server was not reached.
This commit reverts the original fix and creates a new function
to check if RM info is open or in err state to avoid
mixing RM is open but actually no info and warning with empty info.

Fixes: a81dd1ee5747 ("common.sh: redmine_bug_is_open(): Consider empty RM_STATUS_ID to be open")
Change-Id: Ib186f9b5a0284176733573902f3e7160c2ce683b
Signed-off-by: Roi Dayan <roid@nvidia.com>