common-testing.sh: Always clear remote if executed commands on remote
Many tests already config remote nic/tunnel and assume cleanup
will clean everything.
The issue is when there is no remote and tests fail for not able to
connect to remote.
Use a global variable to check if test executed commands on remote
and only if so then try to clean remote.

Change-Id: Ib3bc16f19640f71ee80022b2cccade935a2b7305
Signed-off-by: Roi Dayan <roid@nvidia.com>