test-ovs-dpdk-meter.sh: Optimize the meter testing
Test is sending burst of buckets for 3 seconds and the
expected packets calculated from the timing and meter
but it should account also for reply packets.
Increase the burst and time of sending the packets
this will show that the number of packets sent is indeed
a lot higher than on the destination netdev including replies.
Also adding 1 second optimize the calculation of expected packets
while the burst is 100 more packets.

Issue: 3746596
Change-Id: Iaf29eee1cef44962dc3113c4d487f4f2974f98d3
Signed-off-by: Roi Dayan <roid@nvidia.com>