common.sh: Fix check_simx_errors() key words
"ERR" string can be part of a debug message which not always
indicate about real error.
For example, when creating some objects, "ERR" string can
be set as a value for a field.
The right string to look for in SimX logs is "ERR-" which
indicate about ERR message.

Issue: None
Signed-off-by: Lior Nahmanson <liorna@nvidia.com>
Change-Id: I81f002da6030f96475d93ce4cc6e6a80026b9022