test-ovs-local-mirror-int-port: Move ignore on dump_create to common
This reverts commit e6c386cddd711e5b4174bec4960db1e2f26c68f3.
Move the ignore to common as it still can reproduce in other tests.
This looks like a race between removing bridge and ports and dump
starting on the tunnel port on the bridge.
This is not reproducing manually so its hard to say if cleaning
the bridge from flows and ports before deleting the bridge can
solve this or maybe the log could probably be a debug instead of err.
This is not an important log so ignore it.

Change-Id: I128f6c23f8712d96d6efc581cdd3a425344ef734
Signed-off-by: Roi Dayan <roid@nvidia.com>