common.sh: Update dump_create expected err msg
In other tests like ovs-ct-geneve-over-vlan-stack-devices-vf-lag,
we face this race between removing bridge/ports and dump starting
on the tunnel port on the bridge.

"netdev_offload_tc(revalidator57)|ERR|dump_create: failed to get
ifindex for geneve0: No such device"

Issue is not reproducing manually so its hard to say if cleaning
the bridge from flows and ports before deleting the bridge can
solve this. This is not an important log so ignore it.

Issue: 4071116
Signed-off-by: Ziyad Atiyyeh <ziyadat@nvidia.com>
Change-Id: Iaf03b98cd90912ff7fe6b26a46f833d32d5268e1