Revert "common-testing: ovs_add_ct_rules(): Add icmp rule based on proto"
Currently some tests comaplains about an error in the log
 "pf0: rte_flow creation failed: 16 (Drop action is mutually-exclusive with any other action, except for Count/Sample/Age action)."

This happens because an ipv6 discovery packet is sent which has no
openflow rule to match it so it get dropped, leading to creating an
offload with decap + drop action which is not supported.

The cited commit removed adding openflow rule with icmp6 match which
caused the above error. Revert it.

[1]: 904d72a637e2 ("common-testing: ovs_add_ct_rules(): Add icmp rule based on proto")

Issue: 3965362
Change-Id: I8caad5bae71191a7f2b4c5cd7b5d20e0f8be06f2
Signed-off-by: Salem Sol <salems@nvidia.com>