common-testing.sh: Don't test for iperf process on remote
The iperf client executed on remote in background and the local pid
of the execution being tested.
At times ssh is slow it could be an ssh to grep the iperf process
could be too late while the remote iperf process already finished
resulting in false error that iperf process failed.
Remove the check. tests should still fail for not having expected
rules and/or traffic.

Change-Id: I0a8c24427287b067cacc04fd7c727fcc77fbf94a
Signed-off-by: Roi Dayan <roid@nvidia.com>