test-all.py: Get dpdk and doca version from ovs-vswitchd
Since openvswitch is compiled static and could be with any
dpdk and doca version then take the versions from what it
reports rather the pc files that may or even may not exists
in the system.

While updating this also refactor the mlnx ofed and prep env html
function names.

Change-Id: Icc2d5d9a9bfab7c4592560fae86758d2d9b5da7c
Signed-off-by: Roi Dayan <roid@nvidia.com>